feat(learner): move floating player to root layout to make it persistent
